refactor: 봉사 일정 Facade Layer command/query 분리
