+++ src/main/java/project/volunteer/domain/scheduleParticipation/api/ScheduleParticipantController.java
@@ -4,7 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import project.volunteer.domain.scheduleParticipation.api.dto.*;
-import project.volunteer.domain.scheduleParticipation.mapper.ScheduleParticipantFacade;
+import project.volunteer.domain.scheduleParticipation.service.ScheduleParticipationCommandFacade;
+import project.volunteer.domain.scheduleParticipation.service.ScheduleParticipationQueryFacade;
 import project.volunteer.domain.scheduleParticipation.service.dto.CancelledParticipantList;
 import project.volunteer.domain.scheduleParticipation.service.dto.CompletedParticipantList;
 import project.volunteer.domain.scheduleParticipation.service.dto.ParticipatingParticipantList;
@@ -18,19 +19,20 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/recruitment")
 public class ScheduleParticipantController {
-    private final ScheduleParticipantFacade scheduleParticipantFacade;
+    private final ScheduleParticipationCommandFacade scheduleParticipationCommandFacade;
+    private final ScheduleParticipationQueryFacade scheduleParticipationQueryFacade;
 
     @OrganizationAuth(auth = OrganizationAuth.Auth.ORGANIZATION_TEAM)
     @PutMapping("/{recruitmentNo}/schedule/{scheduleNo}/join")
     public ResponseEntity scheduleParticipation(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo){
-        scheduleParticipantFacade.participateVolunteerPostSchedule(SecurityUtil.getLoginUserNo(), recruitmentNo, scheduleNo);
+        scheduleParticipationCommandFacade.participateSchedule(SecurityUtil.getLoginUserNo(), recruitmentNo, scheduleNo);
         return ResponseEntity.ok().build();
     }
 
     @OrganizationAuth(auth = OrganizationAuth.Auth.ORGANIZATION_TEAM)
     @PutMapping("/{recruitmentNo}/schedule/{scheduleNo}/cancel")
     public ResponseEntity scheduleCancelRequest(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo){
-        scheduleParticipantFacade.cancelParticipationVolunteerPostSchedule(SecurityUtil.getLoginUserNo(), recruitmentNo, scheduleNo);
+        scheduleParticipationCommandFacade.cancelParticipationVolunteerPostSchedule(SecurityUtil.getLoginUserNo(), recruitmentNo, scheduleNo);
         return ResponseEntity.ok().build();
     }
 
@@ -38,7 +40,7 @@ public class ScheduleParticipantController {
     @PutMapping("/{recruitmentNo}/schedule/{scheduleNo}/cancelling")
     public ResponseEntity scheduleCancelApproval(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo,
                                                  @RequestBody @Valid CancelApproval dto){
-        scheduleParticipantFacade.approvalCancellationVolunteerPostSchedule(scheduleNo, dto.getNo());
+        scheduleParticipationCommandFacade.approvalCancellationVolunteerPostSchedule(scheduleNo, dto.getNo());
         return ResponseEntity.ok().build();
     }
 
@@ -47,7 +49,7 @@ public class ScheduleParticipantController {
     public ResponseEntity scheduleCompleteApproval(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo,
                                                    @RequestBody @Valid CompleteApproval dto){
 
-        scheduleParticipantFacade.approvalCompletionVolunteerPostSchedule(scheduleNo, dto.getCompletedList());
+        scheduleParticipationCommandFacade.approvalCompletionVolunteerPostSchedule(scheduleNo, dto.getCompletedList());
         return ResponseEntity.ok().build();
     }
 
@@ -55,7 +57,7 @@ public class ScheduleParticipantController {
     @GetMapping("/{recruitmentNo}/schedule/{scheduleNo}/participating")
     public ResponseEntity<ParticipatingParticipantListResponse> scheduleParticipantList(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo){
 
-        List<ParticipatingParticipantList> participating = scheduleParticipantFacade.findParticipatingParticipantsSchedule(scheduleNo);
+        List<ParticipatingParticipantList> participating = scheduleParticipationQueryFacade.findParticipatingParticipantsSchedule(scheduleNo);
         return ResponseEntity.ok(new ParticipatingParticipantListResponse(participating));
     }
 
@@ -63,7 +65,7 @@ public class ScheduleParticipantController {
     @GetMapping("/{recruitmentNo}/schedule/{scheduleNo}/cancelling")
     public ResponseEntity<CancelledParticipantListResponse> scheduleCancelledParticipantList(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo){
 
-        List<CancelledParticipantList> cancellingParticipants = scheduleParticipantFacade.findCancelledParticipantsSchedule(scheduleNo);
+        List<CancelledParticipantList> cancellingParticipants = scheduleParticipationQueryFacade.findCancelledParticipantsSchedule(scheduleNo);
         return ResponseEntity.ok(new CancelledParticipantListResponse(cancellingParticipants));
     }
 
@@ -71,7 +73,7 @@ public class ScheduleParticipantController {
     @GetMapping("/{recruitmentNo}/schedule/{scheduleNo}/completion")
     public ResponseEntity<CompletedParticipantListResponse> scheduleCompletedParticipantList(@PathVariable Long recruitmentNo, @PathVariable Long scheduleNo){
 
-        List<CompletedParticipantList> completedParticipants = scheduleParticipantFacade.findCompletedParticipantsSchedule(scheduleNo);
+        List<CompletedParticipantList> completedParticipants = scheduleParticipationQueryFacade.findCompletedParticipantsSchedule(scheduleNo);
         return ResponseEntity.ok(new CompletedParticipantListResponse(completedParticipants));
     }
 }
